HODT added and regular insertion done.
The insertion on this class is simple. Split face in three if point lies inside face, split in four faces if poit lies on edge, and add correctly the point if it lies outside the convex hull. For dimensions lower than 2, all the process is inhereted and used from Delaunay Triangulation.